Basic funtionality for Texture changing implemented
