--- Assets/Scripts/Systems/Painting System/ColorManipulator.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: a70f9be50bfb11c4d9288bde4b2b2ba6
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
