--- Assets/Scripts/Systems/Painting System/TextureManipulator.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 3c2ba392902c0a74ab2b0e75aeb6d17c
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
