+++ Assets/UI Assets/RIght Menu/Source Surface Textures.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: de516ce9a2f240a498952c779016fc47
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
